Move config to build and change its structure
