--- packages/frontend/src/pages/config.ts
@@ -1,14 +1,0 @@
-export const config = {
-  twitterLink: 'https://twitter.com/l2beat',
-  discordLink: 'https://discord.gg/eaVKXPmtWk',
-  githubLink: 'https://github.com/l2beat/l2beat',
-  youTubeLink:
-    'https://www.youtube.com/channel/UCDrl-fNXFjOoykr4lQij9BA/videos',
-  mediumLink: 'https://medium.com/l2beat',
-  forumLink: 'https://gov.l2beat.com/',
-  showBanner: Date.now() < new Date('2022-09-23').getTime(),
-  showGitcoinOption: true,
-  showPermissionedAccounts: true,
-  showBridgePages: false,
-  showActivity: false,
-}
